requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytest
pytest-asyncio
